Moved key getters to EntityConverter and added unit test
